move OsVersion to its own toplevel module
This is required to avoid cyclic imports between bci_build.package and
bci_build.registry.
