+++ src/bci_build/package/appcontainers.py
@@ -2,13 +2,13 @@
 
 from pathlib import Path
 
-from bci_build.package import ALL_NONBASE_OS_VERSIONS
-from bci_build.package import CAN_BE_LATEST_OS_VERSION
+from bci_build.osversion import ALL_NONBASE_OS_VERSIONS
+from bci_build.osversion import CAN_BE_LATEST_OS_VERSION
+from bci_build.osversion import OsVersion
 from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import ApplicationStackContainer
 from bci_build.package import BuildType
 from bci_build.package import OsContainer
-from bci_build.package import OsVersion
 from bci_build.package import Package
 from bci_build.package import PackageType
 from bci_build.package import ParseVersion
