+++ src/bci_build/package/base.py
@@ -5,10 +5,10 @@ from pathlib import Path
 
 from jinja2 import Template
 
+from bci_build.osversion import OsVersion
 from bci_build.package import Arch
 from bci_build.package import BuildType
 from bci_build.package import OsContainer
-from bci_build.package import OsVersion
 from bci_build.package import Package
 from bci_build.package import PackageType
 from bci_build.package import SupportLevel
