+++ src/bci_build/package/mariadb.py
@@ -3,12 +3,12 @@
 import re
 from pathlib import Path
 
-from bci_build.package import ALL_NONBASE_OS_VERSIONS
-from bci_build.package import CAN_BE_LATEST_OS_VERSION
+from bci_build.osversion import ALL_NONBASE_OS_VERSIONS
+from bci_build.osversion import CAN_BE_LATEST_OS_VERSION
+from bci_build.osversion import OsVersion
 from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import ApplicationStackContainer
 from bci_build.package import BuildType
-from bci_build.package import OsVersion
 from bci_build.package import ParseVersion
 from bci_build.package import Replacement
 from bci_build.package import SupportLevel
