+++ src/bci_build/package/node.py
@@ -3,10 +3,10 @@
 import datetime
 from typing import Literal
 
-from bci_build.package import CAN_BE_LATEST_OS_VERSION
-from bci_build.package import _SUPPORTED_UNTIL_SLE
+from bci_build.osversion import CAN_BE_LATEST_OS_VERSION
+from bci_build.osversion import _SUPPORTED_UNTIL_SLE
+from bci_build.osversion import OsVersion
 from bci_build.package import DevelopmentContainer
-from bci_build.package import OsVersion
 from bci_build.package import Replacement
 from bci_build.package import SupportLevel
