+++ src/bci_build/package/openjdk.py
@@ -4,10 +4,10 @@ import os
 from itertools import product
 from typing import Literal
 
+from bci_build.osversion import OsVersion
 from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import Arch
 from bci_build.package import DevelopmentContainer
-from bci_build.package import OsVersion
 from bci_build.package import Replacement
 from bci_build.package import SupportLevel
 from bci_build.package import _build_tag_prefix
