+++ src/bci_build/package/rust.py
@@ -3,8 +3,8 @@
 import datetime
 from itertools import product
 
-from bci_build.package import ALL_NONBASE_OS_VERSIONS
-from bci_build.package import CAN_BE_LATEST_OS_VERSION
+from bci_build.osversion import ALL_NONBASE_OS_VERSIONS
+from bci_build.osversion import CAN_BE_LATEST_OS_VERSION
 from bci_build.package import DevelopmentContainer
 from bci_build.package import Replacement
 from bci_build.package import SupportLevel
