+++ src/bci_build/package/spack.py
@@ -1,11 +1,11 @@
 """HPC container for Spack package manager."""
 
-from bci_build.package import CAN_BE_LATEST_OS_VERSION
+from bci_build.osversion import CAN_BE_LATEST_OS_VERSION
+from bci_build.osversion import _SUPPORTED_UNTIL_SLE
+from bci_build.osversion import OsVersion
 from bci_build.package import DOCKERFILE_RUN
-from bci_build.package import _SUPPORTED_UNTIL_SLE
 from bci_build.package import Arch
 from bci_build.package import DevelopmentContainer
-from bci_build.package import OsVersion
 from bci_build.package import ParseVersion
 from bci_build.package import SupportLevel
 from bci_build.package import generate_disk_size_constraints
