+++ tests/test_crate.py
@@ -1,7 +1,7 @@
 from bci_build.containercrate import ContainerCrate
+from bci_build.osversion import OsVersion
 from bci_build.package import BuildType
 from bci_build.package import DevelopmentContainer
-from bci_build.package import OsVersion
 
 _BASE_KWARGS = {
     "name": "test",
